removed unneeded files and combined the logic
